# core/docs/swagger.yaml
definitions:
...
  apprun_pkg_response.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  apprun_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      error:
        $ref: '#/definitions/apprun_pkg_response.ErrorInfo'
      message:
        type: string
      request_id:
        type: string
      success:
        type: boolean
    type: object
...
  modules_config.GetConfigResponse:
    properties:
      is_dynamic:
        description: Whether it's a dynamic configuration
        example: false
        type: boolean
      key:
        description: Configuration key
        example: app.name
        type: string
      source:
        description: 'Source: "database", "file", "env", "default"'
        example: default
        type: string
      value:
        description: Configuration value
        example: apprun
        type: string
    type: object
  modules_config.ListConfigsResponse:
    properties:
      configs:
        additionalProperties:
          type: string
        description: Key-value mapping of dynamic configurations
        type: object
      count:
        description: Number of configuration items
        example: 3
        type: integer
    type: object
  modules_config.UpdateConfigRequest:
    properties:
      key:
        description: Configuration key
        example: poc.enabled
        type: string
      value:
        description: New value
        example: "true"
        type: string
    required:
    - key
    - value
    type: object
  modules_config.UpdateConfigResponse:
    properties:
      key:
        example: poc.enabled
        type: string
      value:
        example: "true"
        type: string
    type: object
info:
  contact:
    email: support@websoft9.com
    name: API Support
  description: AppRun Platform REST API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: AppRun API
  version: "1.0"
paths:
...
  /api/config:
    delete:
      consumes:
      - application/json
      description: |-
        Delete a dynamic configuration item from database (config will fallback to file or default value).
        Only dynamic configurations (db:true) can be deleted.
        After deletion, the config will use the value from config files or built-in defaults.
      parameters:
      - description: Configuration key
        example: poc.enabled
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deletion successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Missing key parameter or deletion failed
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
      summary: Delete configuration item
      tags:
      - config
    get:
      consumes:
      - application/json
      description: Query a single configuration item by key, returns value, source
        and dynamic flag
      parameters:
      - description: Configuration key, e.g. app.name
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Configuration retrieved successfully
          schema:
            $ref: '#/definitions/modules_config.GetConfigResponse'
        "400":
          description: Missing key parameter
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "404":
          description: Configuration not found
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
      summary: Get configuration item
      tags:
      - config
    put:
      consumes:
      - application/json
      description: |-
        Update a single dynamic configuration item (only for db:true configs).
        Static configurations (db:false) cannot be updated via API.
        Changes are persisted to database and take effect immediately.
      parameters:
      - description: Configuration update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/modules_config.UpdateConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Configuration updated successfully
          schema:
            $ref: '#/definitions/modules_config.UpdateConfigResponse'
        "400":
          description: Invalid request or config not allowed to store in database
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
      summary: Update configuration item
      tags:
      - config
...
  /api/config/list:
    get:
      consumes:
      - application/json
      description: |-
        Returns all dynamic configuration items stored in database.
        This does not include static configurations from files.
        Use this to see which configs have been overridden dynamically.
      produces:
      - application/json
      responses:
        "200":
          description: Configuration list
          schema:
            $ref: '#/definitions/modules_config.ListConfigsResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
      summary: List dynamic configurations
      tags:
      - config
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"